+++ Cpufit/lm_fit.cpp
@@ -7,15 +7,15 @@
 #include <numeric>
 
 LMFit::LMFit(
-    double const * const data,
-    double const * const weights,
+    float const * const data,
+    float const * const weights,
     Info const & info,
-    double const * const initial_parameters,
+    float const * const initial_parameters,
     int const * const parameters_to_fit,
     char * const user_info,
-    double * output_parameters,
+    float * output_parameters,
     int * output_states,
-    double * output_chi_squares,
+    float * output_chi_squares,
     int * output_n_iterations
     ) :
     data_(data),
@@ -34,7 +34,7 @@ LMFit::~LMFit()
 {
 }
 
-void LMFit::run(double const tolerance)
+void LMFit::run(float const tolerance)
 {
     for (std::size_t fit_index = 0; fit_index < info_.n_fits_; fit_index++)
     {
