+++ Cpufit/lm_fit.h
@@ -9,31 +9,31 @@ class LMFit
 {
 public:
     LMFit(
-        double const * data,
-        double const * weights,
+        float const * data,
+        float const * weights,
         Info const& info,
-        double const * initial_parameters,
+        float const * initial_parameters,
         int const * parameters_to_fit,
         char * user_info,
-        double * output_parameters,
+        float * output_parameters,
         int * output_states,
-        double * output_chi_squares,
+        float * output_chi_squares,
         int * output_n_iterations);
 
     virtual ~LMFit();
 
-    void run(double const tolerance);
+    void run(float const tolerance);
         
 private:
-    double const * const data_;
-    double const * const weights_;
-    double const * const initial_parameters_;
+    float const * const data_;
+    float const * const weights_;
+    float const * const initial_parameters_;
     int const * const parameters_to_fit_;
     char * const user_info_;
 
-    double * output_parameters_;
+    float * output_parameters_;
     int * output_states_;
-    double * output_chi_squares_;
+    float * output_chi_squares_;
     int * output_n_iterations_;
 
     Info const & info_;
@@ -43,17 +43,17 @@ class LMFitCPP
 {
 public:
     LMFitCPP(
-        double const tolerance,
+        float const tolerance,
         std::size_t const fit_index,
-        double const * data,
-        double const * weight,
+        float const * data,
+        float const * weight,
         Info const & info,
-        double const * initial_parameters,
+        float const * initial_parameters,
         int const * parameters_to_fit,
         char * user_info,
-        double * output_parameters,
+        float * output_parameters,
         int * output_states,
-        double * output_chi_squares,
+        float * output_chi_squares,
         int * output_n_iterations);
 
     virtual ~LMFitCPP()
@@ -65,43 +65,43 @@ private:
 	void calc_model();
     void calc_coefficients();
 
-    void calc_curve_values(std::vector<double>& curve, std::vector<double>& derivatives);
+    void calc_curve_values(std::vector<float>& curve, std::vector<float>& derivatives);
 
-    void calc_values_gauss2d(std::vector<double>& gaussian);
-    void calc_derivatives_gauss2d(std::vector<double> & derivatives);
+    void calc_values_gauss2d(std::vector<float>& gaussian);
+    void calc_derivatives_gauss2d(std::vector<float> & derivatives);
 
-    void calc_values_gauss2delliptic(std::vector<double>& gaussian);
-    void calc_derivatives_gauss2delliptic(std::vector<double> & derivatives);
+    void calc_values_gauss2delliptic(std::vector<float>& gaussian);
+    void calc_derivatives_gauss2delliptic(std::vector<float> & derivatives);
 
-    void calc_values_gauss2drotated(std::vector<double>& gaussian);
-    void calc_derivatives_gauss2drotated(std::vector<double> & derivatives);
+    void calc_values_gauss2drotated(std::vector<float>& gaussian);
+    void calc_derivatives_gauss2drotated(std::vector<float> & derivatives);
 
-    void calc_values_gauss1d(std::vector<double>& gaussian);
-    void calc_derivatives_gauss1d(std::vector<double> & derivatives);
+    void calc_values_gauss1d(std::vector<float>& gaussian);
+    void calc_derivatives_gauss1d(std::vector<float> & derivatives);
 
-    void calc_values_cauchy2delliptic(std::vector<double>& cauchy);
-    void calc_derivatives_cauchy2delliptic(std::vector<double> & derivatives);
+    void calc_values_cauchy2delliptic(std::vector<float>& cauchy);
+    void calc_derivatives_cauchy2delliptic(std::vector<float> & derivatives);
 
-    void calc_values_linear1d(std::vector<double>& line);
-    void calc_derivatives_linear1d(std::vector<double> & derivatives);
+    void calc_values_linear1d(std::vector<float>& line);
+    void calc_derivatives_linear1d(std::vector<float> & derivatives);
 
-    void calc_values_fletcher_powell_helix(std::vector<double>& values);
-    void calc_derivatives_fletcher_powell_helix(std::vector<double> & derivatives);
+    void calc_values_fletcher_powell_helix(std::vector<float>& values);
+    void calc_derivatives_fletcher_powell_helix(std::vector<float> & derivatives);
 
-    void calc_values_brown_dennis(std::vector<double>& values);
-    void calc_derivatives_brown_dennis(std::vector<double> & derivatives);
+    void calc_values_brown_dennis(std::vector<float>& values);
+    void calc_derivatives_brown_dennis(std::vector<float> & derivatives);
 
-    void calc_values_ramsey_var_p(std::vector<double>& values);
-    void calc_derivatives_ramsey_var_p(std::vector<double> & derivatives);
+    void calc_values_ramsey_var_p(std::vector<float>& values);
+    void calc_derivatives_ramsey_var_p(std::vector<float> & derivatives);
 
-    void calculate_hessian(std::vector<double> const & derivatives,
-        std::vector<double> const & curve);
+    void calculate_hessian(std::vector<float> const & derivatives,
+        std::vector<float> const & curve);
 
-    void calc_gradient(std::vector<double> const & derivatives,
-        std::vector<double> const & curve);
+    void calc_gradient(std::vector<float> const & derivatives,
+        std::vector<float> const & curve);
 
     void calc_chi_square(
-        std::vector<double> const & curve);
+        std::vector<float> const & curve);
 
     template< class T >
     void decompose_hessian_LUP(std::vector<T> & decomposed_hessian, std::vector<T> const & hessian);
@@ -125,44 +125,44 @@ public:
 private:
 
     std::size_t const fit_index_;
-    double const * const data_;
-    double const * const weight_;
-    double const * const initial_parameters_;
+    float const * const data_;
+    float const * const weight_;
+    float const * const initial_parameters_;
     int const * const parameters_to_fit_;
 
     bool converged_;
-    double * parameters_;
+    float * parameters_;
     int * state_;
-    double * chi_square_;
+    float * chi_square_;
     int * n_iterations_;
 
-    std::vector<double> prev_parameters_;
+    std::vector<float> prev_parameters_;
     Info const & info_;
 
-    double lambda_;
-    double lambda_lower_bound_;
-    double lambda_upper_bound_;
-    double step_bound_;
-    double actual_reduction_;
-    double predicted_reduction_;
-    double directive_derivative_;
-    double approximation_ratio_;
-    double phi_;
-    double phi_derivative_;
+    float lambda_;
+    float lambda_lower_bound_;
+    float lambda_upper_bound_;
+    float step_bound_;
+    float actual_reduction_;
+    float predicted_reduction_;
+    float directive_derivative_;
+    float approximation_ratio_;
+    float phi_;
+    float phi_derivative_;
 
-    std::vector<double> curve_;
-    std::vector<double> derivatives_;
-    std::vector<double> temp_derivatives_;
-    std::vector<double> hessian_;
-    std::vector<double> decomposed_hessian_;
-    std::vector<double> inverted_hessian_;
+    std::vector<float> curve_;
+    std::vector<float> derivatives_;
+    std::vector<float> temp_derivatives_;
+    std::vector<float> hessian_;
+    std::vector<float> decomposed_hessian_;
+    std::vector<float> inverted_hessian_;
     std::vector<int> pivot_array_;
-    std::vector<double> modified_hessian_;
-    std::vector<double> gradient_;
-    std::vector<double> delta_;
-    std::vector<double> scaling_vector_;
-    double prev_chi_square_;
-    double const tolerance_;
+    std::vector<float> modified_hessian_;
+    std::vector<float> gradient_;
+    std::vector<float> delta_;
+    std::vector<float> scaling_vector_;
+    float prev_chi_square_;
+    float const tolerance_;
 
     char * const user_info_;
 };
